move common adapter logic to adapter utils
