+++ backend/adapter_utils.go
@@ -1,0 +1,39 @@
+package backend
+
+import (
+	"context"
+	"net/http"
+
+	"github.com/grafana/grafana-plugin-sdk-go/backend/httpclient"
+)
+
+func setupContext(ctx context.Context, endpoint Endpoint) context.Context {
+	ctx = WithEndpoint(ctx, endpoint)
+	ctx = propagateTenantIDIfPresent(ctx)
+
+	return ctx
+}
+
+func withHeaderMiddleware(ctx context.Context, headers http.Header) context.Context {
+	if len(headers) > 0 {
+		ctx = httpclient.WithContextualMiddleware(ctx,
+			httpclient.MiddlewareFunc(func(opts httpclient.Options, next http.RoundTripper) http.RoundTripper {
+				if !opts.ForwardHTTPHeaders {
+					return next
+				}
+
+				return httpclient.RoundTripperFunc(func(qreq *http.Request) (*http.Response, error) {
+					// Only set a header if it is not already set.
+					for k, v := range headers {
+						if qreq.Header.Get(k) == "" {
+							for _, vv := range v {
+								qreq.Header.Add(k, vv)
+							}
+						}
+					}
+					return next.RoundTrip(qreq)
+				})
+			}))
+	}
+	return ctx
+}
