+++ backend/data_adapter.go
@@ -2,9 +2,7 @@ package backend
 
 import (
 	"context"
-	"net/http"
 
-	"github.com/grafana/grafana-plugin-sdk-go/backend/httpclient"
 	"github.com/grafana/grafana-plugin-sdk-go/genproto/pluginv2"
 )
 
@@ -19,35 +17,10 @@ func newDataSDKAdapter(handler QueryDataHandler) *dataSDKAdapter {
 	}
 }
 
-func withHeaderMiddleware(ctx context.Context, headers http.Header) context.Context {
-	if len(headers) > 0 {
-		ctx = httpclient.WithContextualMiddleware(ctx,
-			httpclient.MiddlewareFunc(func(opts httpclient.Options, next http.RoundTripper) http.RoundTripper {
-				if !opts.ForwardHTTPHeaders {
-					return next
-				}
-
-				return httpclient.RoundTripperFunc(func(qreq *http.Request) (*http.Response, error) {
-					// Only set a header if it is not already set.
-					for k, v := range headers {
-						if qreq.Header.Get(k) == "" {
-							for _, vv := range v {
-								qreq.Header.Add(k, vv)
-							}
-						}
-					}
-					return next.RoundTrip(qreq)
-				})
-			}))
-	}
-	return ctx
-}
-
 func (a *dataSDKAdapter) QueryData(ctx context.Context, req *pluginv2.QueryDataRequest) (*pluginv2.QueryDataResponse, error) {
-	ctx = WithEndpoint(ctx, EndpointQueryData)
-	ctx = propagateTenantIDIfPresent(ctx)
-	ctx = WithGrafanaConfig(ctx, NewGrafanaCfg(req.PluginContext.GrafanaConfig))
+	ctx = setupContext(ctx, EndpointQueryData)
 	parsedReq := FromProto().QueryDataRequest(req)
+	ctx = WithGrafanaConfig(ctx, parsedReq.PluginContext.GrafanaConfig)
 	ctx = WithPluginContext(ctx, parsedReq.PluginContext)
 	ctx = WithUser(ctx, parsedReq.PluginContext.User)
 	ctx = withHeaderMiddleware(ctx, parsedReq.GetHTTPHeaders())
