+++ backend/diagnostics_adapter.go
@@ -45,11 +45,11 @@ func (a *diagnosticsSDKAdapter) CollectMetrics(_ context.Context, _ *pluginv2.Co
 }
 
 func (a *diagnosticsSDKAdapter) CheckHealth(ctx context.Context, protoReq *pluginv2.CheckHealthRequest) (*pluginv2.CheckHealthResponse, error) {
+	ctx = setupContext(ctx, EndpointCheckHealth)
+
 	if a.checkHealthHandler != nil {
-		ctx = WithEndpoint(ctx, EndpointCheckHealth)
-		ctx = propagateTenantIDIfPresent(ctx)
-		ctx = WithGrafanaConfig(ctx, NewGrafanaCfg(protoReq.PluginContext.GrafanaConfig))
 		parsedReq := FromProto().CheckHealthRequest(protoReq)
+		ctx = WithGrafanaConfig(ctx, parsedReq.PluginContext.GrafanaConfig)
 		ctx = WithPluginContext(ctx, parsedReq.PluginContext)
 		ctx = WithUser(ctx, parsedReq.PluginContext.User)
 		ctx = withHeaderMiddleware(ctx, parsedReq.GetHTTPHeaders())
